machine-learning/requirements.txt
autopep8==1.5.4
awscli==1.18.203
botocore==1.19.43
CacheControl==0.12.6
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.3
cvxpy==1.1.7
cycler==0.10.0
docopt==0.6.2
docutils==0.15.2
ecos==2.0.7.post1
firebase-admin==4.5.1
Flask==1.1.2
google-api-core==1.25.0
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-cloud-core==1.5.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.35.0
google-crc32c==1.1.1
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.34.1
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
msgpack==1.0.2
numpy==1.19.5
osqp==0.6.2.post0
pandas==1.2.0
Pillow==8.1.0
pipreqs==0.4.10
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.3.1
qdldl==0.1.5.post0
requests==2.25.1
rsa==4.5
s3transfer==0.3.3
scikit-learn==0.24.0
scipy==1.6.0
scs==2.1.2
seaborn==0.11.1
six==1.15.0
threadpoolctl==2.1.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
Werkzeug==1.0.1
yarg==0.1.9
